+++ src/resources/views/invitation/create.blade.php
@@ -1,25 +1,25 @@
 @extends('appshell::layouts.private')
 
 @section('title')
-    {{ __('Create new user') }}
+    {{ __('Invite new user') }}
 @stop
 
 @section('content')
 
-@include('appshell::user._subnav', ['active' => 'users'])
+@include('appshell::user._subnav', ['active' => 'invitations'])
 
-{!! Form::model($user, ['route' => 'appshell.user.store', 'autocomplete' => 'off', 'class' => 'row']) !!}
+{!! Form::model($invitation, ['route' => 'appshell.invitation.store', 'autocomplete' => 'off', 'class' => 'row']) !!}
 
 <div class="col-12 col-md-6 col-lg-8 col-xl-9">
     <div class="card card-accent-success">
-        <div class="card-header">{{ __('Enter Account Details') }}</div>
 
+        <div class="card-header">{{ __('Invitee Details') }}</div>
         <div class="card-body">
-            @include('appshell::user._form')
+            @include('appshell::invitation._form')
         </div>
 
         <div class="card-footer">
-            <button class="btn btn-success">{{ __('Create user') }}</button>
+            <button class="btn btn-success">{{ __('Create invitation') }}</button>
             <a href="#" onclick="history.back();" class="btn btn-link text-muted">{{ __('Cancel') }}</a>
         </div>
     </div>
@@ -30,7 +30,7 @@
 
         <div class="card-header">{{ __('Roles') }}</div>
         <div class="card-body">
-            @include('appshell::role._assignment', ['model' => $user])
+            @include('appshell::role._assignment', ['model' => $invitation])
         </div>
     </div>
 </div>